Adjust configuration of "Arduino Lint" workflow following Library Manager submission
The "Arduino Lint" GitHub Actions workflow checks for problems with the Arduino metadata and structure of the library.

There are different requirements depending on whether the library has been submitted to Arduino Library Manager
registry. Previously the workflow was configured as appropriate for a pre-submission library.  The library has since
been submitted, which made the configuration inappropriate and caused spurious workflow run failures.